changed classes.py layout
